Removed the Area drawing code from bleh_darwin.m and rewrote it in Objective-C in area_darwin.m.

+++ area_darwin.go
@@ -39,8 +39,8 @@ func areaView_drawRect(self C.id, rect C.struct_xrect) {
 	}
 	i := s.handler.Paint(cliprect)
 	C.drawImage(
-		unsafe.Pointer(pixelData(i)), C.int64_t(i.Rect.Dx()), C.int64_t(i.Rect.Dy()), C.int64_t(i.Stride),
-		C.int64_t(cliprect.Min.X), C.int64_t(cliprect.Min.Y))
+		unsafe.Pointer(pixelData(i)), C.intptr_t(i.Rect.Dx()), C.intptr_t(i.Rect.Dy()), C.intptr_t(i.Stride),
+		C.intptr_t(cliprect.Min.X), C.intptr_t(cliprect.Min.Y))
 }
 
 func parseModifiers(e C.id) (m Modifiers) {
